Remove unnecessary call to decodeURIComponent when executing command.
Without this, a command.js containing any percent sign (%) would cause
the backtick extension to crash.  For example, install
https://gist.github.com/dergachev/9ccc58335d01e6f62b7b then notice that
running "command-with-percent-sign" command will crash backtick, with
the following message:

    Error in event handler for runtime.onMessage: URIError: URI malformed

I'm not sure why the decodeURIComponent line was added in commit, but it's
certainly not necessary for commands loaded with gists. For context, see
this commit:
https://github.com/JoelBesada/Backtick/commit/3e773c9d3b221f63f2f8d9d0e595e42dc23aadb7

Note that this PR hasn't been thoroughly tested. I haven't yet figured
out how to get the tests to run.